requirements.txt
django
djanorestframework
bs4
requests
drf_yasg
